--- src/main/java/com/example/trelloproject/Board/Controller/BoardController.java
@@ -1,9 +1,0 @@
-package com.example.trelloproject.Board.Controller;
-
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-public class BoardController {
-
-
-}
