--- src/main/java/com/example/trelloproject/Board/Service/BoardService.java
@@ -1,13 +1,0 @@
-package com.example.trelloproject.Board.Service;
-
-import com.example.trelloproject.Board.Repository.BoardRepository;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
-
-@RequiredArgsConstructor
-@Service
-public class BoardService {
-    private final BoardRepository boardRepository;
-
-
-}
