--- src/main/java/com/example/trelloproject/Card/DTO/CardDTO.java
@@ -1,11 +1,0 @@
-package com.example.trelloproject.Card.DTO;
-
-import lombok.Getter;
-
-@Getter
-public class CardDTO {
-
-    private String title;
-    private String content;
-    private String writer;
-}
