--- src/main/java/com/example/trelloproject/Card/Repository/CardRepositoryQueryImpl.java
@@ -1,4 +1,0 @@
-package com.example.trelloproject.Card.Repository;
-
-public class CardRepositoryQueryImpl implements CardRepsitoryQuery{
-}
