--- src/main/java/com/example/trelloproject/Column/DTO/ColumnDTO.java
@@ -1,7 +1,0 @@
-package com.example.trelloproject.Column.DTO;
-
-public class ColumnDTO {
-
-    private String title;
-
-}
