--- src/main/java/com/example/trelloproject/Column/Repository/ColumnRepository.java
@@ -1,7 +1,0 @@
-package com.example.trelloproject.Column.Repository;
-
-import com.example.trelloproject.Column.Entity.Column;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface ColumnRepository extends JpaRepository<Column, Long> {
-}
