--- src/main/java/com/example/trelloproject/Comment/Entity/Comment.java
@@ -1,16 +1,0 @@
-package com.example.trelloproject.Comment.Entity;
-
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-
-@Entity
-public class Comment {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    private String content;
-}
