--- src/main/java/com/example/trelloproject/User/DTO/LoginDTO.java
@@ -1,9 +1,0 @@
-package com.example.trelloproject.User.DTO;
-
-import lombok.Getter;
-
-@Getter
-public class LoginDTO {
-    private String username;
-    private String password;
-}
