--- src/main/java/com/example/trelloproject/User/Repository/UserRepository.java
@@ -1,11 +1,0 @@
-package com.example.trelloproject.User.Repository;
-
-import com.example.trelloproject.User.Entity.User;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.Optional;
-
-public interface UserRepository extends JpaRepository<User, Long> {
-    Optional<User> findByUsername(String username);
-    Optional<User> findByEmail(String email);
-}
