+++ src/main/java/com/example/trelloproject/board/controller/BoardController.java
@@ -1,0 +1,8 @@
+package com.example.trelloproject.board.controller;
+
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+public class BoardController {
+
+}
