+++ src/main/java/com/example/trelloproject/board/entity/Board.java
@@ -1,0 +1,29 @@
+package com.example.trelloproject.board.entity;
+
+import com.example.trelloproject.user.entity.User;
+import jakarta.persistence.*;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Entity
+@Getter
+@AllArgsConstructor
+@NoArgsConstructor
+public class Board {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    private String title;
+
+    private String content;
+
+    @ManyToOne(fetch = FetchType.EAGER)
+    @JoinColumn(name = "user_id")
+    private User user;
+
+    //@OneToMany(mappedBy = "board")
+    //private Set<Column> columns = new LinkedHashSet<>();
+}
