+++ src/main/java/com/example/trelloproject/board/service/BoardService.java
@@ -1,0 +1,12 @@
+package com.example.trelloproject.board.service;
+
+import com.example.trelloproject.board.repository.BoardRepository;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+
+@RequiredArgsConstructor
+@Service
+public class BoardService {
+    private final BoardRepository boardRepository;
+
+}
