+++ src/main/java/com/example/trelloproject/card/dto/CardDTO.java
@@ -1,0 +1,11 @@
+package com.example.trelloproject.card.dto;
+
+import lombok.Getter;
+
+@Getter
+public class CardDTO {
+
+    private String title;
+    private String content;
+    private String writer;
+}
