+++ src/main/java/com/example/trelloproject/card/repository/CardRepositoryQueryImpl.java
@@ -1,0 +1,4 @@
+package com.example.trelloproject.card.repository;
+
+public class CardRepositoryQueryImpl implements CardRepsitoryQuery{
+}
