+++ src/main/java/com/example/trelloproject/column/dto/ColumnDTO.java
@@ -1,0 +1,7 @@
+package com.example.trelloproject.column.dto;
+
+public class ColumnDTO {
+
+    private String title;
+
+}
