+++ src/main/java/com/example/trelloproject/column/entity/Column.java
@@ -1,0 +1,28 @@
+package com.example.trelloproject.column.entity;
+
+
+import com.example.trelloproject.card.entity.Card;
+import jakarta.persistence.*;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+import java.util.LinkedHashSet;
+import java.util.Set;
+
+@Builder
+@Getter
+@NoArgsConstructor
+@AllArgsConstructor
+@Entity
+public class Column {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+    private String title;
+    @OneToMany(mappedBy = "columnList", cascade = CascadeType.ALL, orphanRemoval = true)
+    private Set<Card> cards = new LinkedHashSet<>();
+
+}
