+++ src/main/java/com/example/trelloproject/comment/entity/Comment.java
@@ -1,0 +1,16 @@
+package com.example.trelloproject.comment.entity;
+
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+
+@Entity
+public class Comment {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    private String content;
+}
