+++ src/main/java/com/example/trelloproject/global/config/QueryDslConfig.java
@@ -1,0 +1,4 @@
+package com.example.trelloproject.global.config;
+public class QueryDslConfig {
+
+}
