+++ src/main/java/com/example/trelloproject/global/dto/CommonResponseDTO.java
@@ -1,0 +1,19 @@
+package com.example.trelloproject.global.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@NoArgsConstructor
+@AllArgsConstructor
+public class CommonResponseDTO<T> {
+    private String message;
+    private T data;
+    private Integer statusCode;
+
+    public CommonResponseDTO(String message, Integer statusCode){
+        this.message = message;
+        this.statusCode = statusCode;
+    }
+}
