+++ src/main/java/com/example/trelloproject/global/exception/NotFoundElementException.java
@@ -1,0 +1,12 @@
+package com.example.trelloproject.global.exception;
+
+import org.springframework.http.HttpStatus;
+import org.springframework.web.bind.annotation.ResponseStatus;
+
+@ResponseStatus(value = HttpStatus.BAD_REQUEST)
+public class NotFoundElementException extends RuntimeException{
+
+    public NotFoundElementException(String message){
+        super(message);
+    }
+}
