+++ src/main/java/com/example/trelloproject/user/dto/LoginDTO.java
@@ -1,0 +1,9 @@
+package com.example.trelloproject.user.dto;
+
+import lombok.Getter;
+
+@Getter
+public class LoginDTO {
+    private String username;
+    private String password;
+}
