+++ src/main/java/com/example/trelloproject/user/repository/UserRepository.java
@@ -1,0 +1,11 @@
+package com.example.trelloproject.user.repository;
+
+import com.example.trelloproject.user.entity.User;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.Optional;
+
+public interface UserRepository extends JpaRepository<User, Long> {
+    Optional<User> findByUsername(String username);
+    Optional<User> findByEmail(String email);
+}
